--- src/components/attribute.ts
@@ -1,5 +1,0 @@
-import { Component } from "./index";
-
-export class Attribute extends Component {
-  public componentType = "attribute";
-}
