--- src/components/dimension.ts
@@ -1,5 +1,0 @@
-import { Component } from "./index";
-
-export class Dimension extends Component {
-  public componentType = "dimension";
-}
