+++ src/components/index.ts
@@ -1,4 +1,1 @@
-export { Component } from "./component";
-export { Attribute } from "./attribute";
-export { Dimension } from "./dimension";
-export { Measure } from "./measure";
+export { Attribute, Component, Dimension, Measure } from "./component";
