--- src/components/measure.ts
@@ -1,5 +1,0 @@
-import {Component} from "./index";
-
-export class Measure extends Component {
-  public componentType = "measure";
-}
