+++ src/expressions/iexpr.ts
@@ -1,0 +1,3 @@
+export interface IExpr {
+  resolve(mapping: Map<string, string>): IExpr;
+}
